requirements.txt
adb_shell==0.4.4
adbutils==2.8.0
altgraph==0.17.4
anyio==4.8.0
apkutils2==1.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.1
cigam==0.0.3
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cryptography==44.0.1
curl_cffi==0.12.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
fonttools==4.59.0
fqdn==1.5.1
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.1
multitasking==0.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.3.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.2
pefile==2023.2.7
piexif==1.1.3
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==6.31.1
psutil==6.1.1
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pycparser==2.22
pyelftools==0.31
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.2.1
python-telegram-bot==21.1
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.31.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
schedule==1.2.2
selenium==4.34.2
Send2Trash==1.8.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.14.1
tzdata==2025.2
uiautomator2==3.2.8
uri-template==1.3.0
urllib3==2.5.0
voila==0.5.8
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
xmltodict==0.14.2
yfinance==0.2.65
